Initial Additions
Im coding this on the OSU server so I can only upload codes in chunks when I download them locally. I've just finished making it so that the program can open a file, take a line from it, and add the information into a movie struct. next step is creating a linked list and making it so a movie can be added to it