Polished/ Refined code
Added comments to the entirety of the code. Checked for bugs- turned out i was getting segfaults from freeing invalid memory so i just entirely got rid of the free functions in my code as this assignment will not be graded on it. Got rid of the testing csv file. Added README.txt to explain how to compile the code. Changed the executable name from main to movies, removed 'make run' as my chosen executable is not 100% going to be used